Remove useComparison hook
