+++ x-pack/plugins/apm/public/components/app/transaction_details/index.tsx
@@ -30,6 +30,7 @@ export function TransactionDetails() {
     rangeTo,
     transactionType: transactionTypeFromUrl,
     comparisonEnabled,
+    comparisonType,
   } = query;
   const { start, end } = useTimeRange({ rangeFrom, rangeTo });
   const apmRouter = useApmRouter();
@@ -73,6 +74,7 @@ export function TransactionDetails() {
           transactionName={transactionName}
           isServerlessContext={isServerless}
           comparisonEnabled={comparisonEnabled}
+          comparisonType={comparisonType}
         />
       </ChartPointerEventContextProvider>
