+++ x-pack/plugins/apm/public/components/app/transaction_overview/index.tsx
@@ -26,6 +26,7 @@ export function TransactionOverview() {
       rangeTo,
       transactionType: transactionTypeFromUrl,
       comparisonEnabled,
+      comparisonType,
     },
   } = useApmParams('/services/{serviceName}/transactions');
 
@@ -68,6 +69,7 @@ export function TransactionOverview() {
         end={end}
         isServerlessContext={isServerless}
         comparisonEnabled={comparisonEnabled}
+        comparisonType={comparisonType}
       />
       <EuiSpacer size="s" />
       <EuiPanel hasBorder={true}>
