+++ x-pack/plugins/apm/public/components/shared/charts/transaction_coldstart_rate_chart/index.tsx
@@ -20,9 +20,13 @@ import { useFetcher } from '../../../../hooks/use_fetcher';
 import { useTheme } from '../../../../hooks/use_theme';
 import { TimeseriesChart } from '../timeseries_chart';
 import { useApmServiceContext } from '../../../../context/apm_service/use_apm_service_context';
+import {
+  getComparisonChartTheme,
+  getTimeRangeComparison,
+} from '../../time_comparison/get_time_range_comparison';
 import { useApmParams } from '../../../../hooks/use_apm_params';
 import { useTimeRange } from '../../../../hooks/use_time_range';
-import { useComparison } from '../../../../hooks/use_comparison';
+import { TimeRangeComparisonType } from '../../../../../common/runtime_types/comparison_type_rt';
 
 function yLabelFormat(y?: number | null) {
   return asPercent(y || 0, 1);
@@ -35,6 +39,7 @@ interface Props {
   environment: string;
   transactionName?: string;
   comparisonEnabled?: boolean;
+  comparisonType?: TimeRangeComparisonType;
 }
 
 type ColdstartRate =
@@ -58,6 +63,7 @@ export function TransactionColdstartRateChart({
   kuery,
   transactionName,
   comparisonEnabled,
+  comparisonType,
 }: Props) {
   const theme = useTheme();
 
@@ -68,8 +74,13 @@ export function TransactionColdstartRateChart({
   const { start, end } = useTimeRange({ rangeFrom, rangeTo });
 
   const { serviceName, transactionType } = useApmServiceContext();
-  const { comparisonChartTheme, comparisonStart, comparisonEnd } =
-    useComparison();
+  const comparisonChartTheme = getComparisonChartTheme();
+  const { comparisonStart, comparisonEnd } = getTimeRangeComparison({
+    start,
+    end,
+    comparisonType,
+    comparisonEnabled,
+  });
 
   const endpoint = transactionName
     ? ('GET /internal/apm/services/{serviceName}/transactions/charts/coldstart_rate_by_transaction_name' as const)
